PR feedback: rename `ImmRef` -> `ImmutableRef`
